Move clear_adjustments method into OrderComponents::Adjustments
