+++ core/spec/models/order_components/adjustments_spec.rb
@@ -123,4 +123,17 @@ describe Spree::OrderComponents::Adjustments do
       end
     end
   end
+
+  context "clear_adjustments" do
+    it "should destroy all previous tax adjustments" do
+      adjustment_1 = stub
+      adjustment_1.should_receive :destroy
+      adjustment_2 = stub
+      adjustment_2.should_receive :destroy
+
+      order.stub_chain :adjustments, :tax => [adjustment_1]
+      order.stub :price_adjustments => [adjustment_2]
+      order.clear_adjustments!
+    end
+  end
 end
